requirements.txt
python-binance==1.0.19
ccxt==4.2.25
websocket-client==1.7.0
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
ta==0.10.2
scikit-learn==1.4.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-telegram-bot==20.7
discord.py==2.3.2
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
schedule==1.2.1
asyncio==3.4.3
aiohttp==3.9.1
websockets==12.0
redis==5.0.1
influxdb-client==1.39.0
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.1

